Add a new muParser function base class.
