+++ source/base/function_parser.cc
@@ -130,7 +130,7 @@ namespace internal
       /**
        * PIMPL for mu::Parser.
        */
-      class Parser : public internal::muParserBase
+      class Parser : public internal::FunctionParser::muParserBase
       {
       public:
         operator mu::Parser &()
@@ -158,7 +158,7 @@ FunctionParser<dim>::init_muparser() const
   // check that we have not already initialized the parser on the
   // current thread, i.e., that the current function is only called
   // once per thread
-  ParserData &data = parser_data.get();
+  internal::FunctionParser::ParserData &data = parser_data.get();
   Assert(data.parsers.size() == 0 && data.vars.size() == 0, ExcInternalError());
 
   // initialize the objects for the current thread
@@ -280,7 +280,7 @@ FunctionParser<dim>::value(const Point<dim> & p,
   AssertIndexRange(component, this->n_components);
 
   // initialize the parser if that hasn't happened yet on the current thread
-  ParserData &data = parser_data.get();
+  internal::FunctionParser::ParserData &data = parser_data.get();
   if (data.vars.size() == 0)
     init_muparser();
 
@@ -327,7 +327,7 @@ FunctionParser<dim>::vector_value(const Point<dim> &p,
 
 
   // initialize the parser if that hasn't happened yet on the current thread
-  ParserData &data = parser_data.get();
+  internal::FunctionParser::ParserData &data = parser_data.get();
   if (data.vars.size() == 0)
     init_muparser();
